+++ ListableUI/Sources/HeaderFooter/HeaderFooter.swift
@@ -41,16 +41,18 @@ public struct HeaderFooter<Content:HeaderFooterContent> : AnyHeaderFooter
     
     public init(
         _ content : Content,
-        sizing : Sizing = .thatFits(.init(.atLeast(.default))),
-        layouts : HeaderFooterLayouts = .init(),
+        sizing : Sizing? = nil,
+        layouts : HeaderFooterLayouts? = nil,
         onTap : OnTap? = nil
     ) {
         assertIsValueType(Content.self)
         
         self.content = content
         
-        self.sizing = sizing
-        self.layouts = layouts
+        let defaults = self.content.defaultHeaderFooterProperties
+        
+        self.sizing = finalValue(from: sizing, defaults.sizing, .thatFits(.init(.atLeast(.default))))
+        self.layouts = finalValue(from: layouts, defaults.layouts, .init())
         
         self.onTap = onTap
         
@@ -108,5 +110,21 @@ extension HeaderFooter : SignpostLoggable
             identifier: self.debuggingIdentifier,
             instanceIdentifier: nil
         )
+    }
+}
+
+
+private func finalValue<Value>(
+    from provided : Value?,
+    _ contentDefault : Value?,
+    _ default : @autoclosure () -> Value
+) -> Value
+{
+    if let value = provided {
+        return value
+    } else if let value = contentDefault {
+        return value
+    } else {
+        return `default`()
     }
 }